Create cpp-jump_search-O(n^(0.5)).cpp
Jump Search is a searching algorithm for sorted arrays. The basic idea is to check fewer elements by jumping ahead by fixed steps or skipping some elements in place of searching all elements.

Algorithm
The steps involved in this algorithm are:
(Block size: B and list size: N; list is sorted in ascending order)

    Step 1: Start from index 1
    Step 2: Jump head by B elements. Current position = Current position + B. If position is out of element list, set current position to last position.
    Step 3: If element at current position < target element, then do Linear Search on element from position current position -B to current position else go to step 2. If current position is last position, go to step 4.
    Step 4: Exit. Element not found.

Consider the list to be of size N and Block size of size B.

In the worst case, we have to do N/B jumps and if the element is not present, we perform B-1 comparisons. Therefore, the total number of comparisons in the worst case will be ((N/B) + B-1). The value of the function ((N/B) + B-1) will be minimum when B = √N.